refactor: split guild uc into separate files
